Last patch

+++ lib/blocs/additional_sum_bloc/buy_together_event.dart
@@ -4,7 +4,7 @@ abstract class BuyTogetherEvent{
 }
 
 class SetAdditionalSumEvent extends BuyTogetherEvent {
-  final double additionalSum;
+  final int additionalSum;
 
   SetAdditionalSumEvent({required this.additionalSum});
 }
